s03: Step 4: fix: Move file deletion functionality to FileUpload component
- Move delete functionality from Chat component to FileUpload component
- Add delete buttons to each file in the Upload Files tab
- Add 'Delete All' button to FileUpload component
- Remove delete buttons from Chat component (file selection only)
- Update FileUpload to use correct FileInfo interface
- Add proper loading states and error handling for deletion
- Improve UI layout with delete buttons next to status badges
